Moved Pawn to Chess module
